Revert "Moving StreamingRandomPatches: localRandomKCombinations(), allKCombinationsInner(), and allKCombinations() into Utils."
This reverts commit 98b356baab3133668d9341c186f262f3cebf03b7.
